Extracted `Repository` class
